hack/tests,test/e2e: separate testing and scaffolding of E2E tests
