+++ test/e2e/_incluster-test-code/main_test.go
@@ -15,11 +15,19 @@
 package e2e
 
 import (
+	"flag"
 	"testing"
 
 	f "github.com/operator-framework/operator-sdk/pkg/test"
 )
 
+type testArgs struct {
+	e2eImageName *string
+}
+
+var args = &testArgs{}
+
 func TestMain(m *testing.M) {
+	args.e2eImageName = flag.String("image", "", "operator image name <repository>:<tag> used to push the image, defaults to none (builds image to local docker repo)")
 	f.MainEntry(m)
 }
